inital updates
