+++ nama/token_similarity.py
@@ -7,147 +7,13 @@ from tqdm import tqdm
 
 from .matcher import Matcher
 from .scoring import score_predicted
-
-
-def ngrams(string, n=2):
-    """
-    Generates all n-grams of the given string.
-
-    Parameters
-    ----------
-    string: str
-        The string from which the n-grams will be generated.
-    n: int, optional
-        The number of characters in each n-gram. Default is 2.
-
-    Yields
-    ------
-    str
-        An n-gram of the string.
-
-    """
-    for i in range(len(string) - n + 1):
-        yield string[i:i + n]
-
-
-def nmgrams(string, n=1, m=3):
-    """
-    Generates all n-grams of the given string where the n-grams are of length between n and m.
-
-    Parameters
-    ----------
-    string: str
-        The string from which the n-grams will be generated.
-    n: int, optional
-        The minimum length of the n-grams to be generated. Default is 1.
-    m: int, optional
-        The maximum length of the n-grams to be generated. Default is 3.
-
-    Yields
-    ------
-    str
-        An n-gram of the string where the length is between n and m.
-
-    """
-    for j in range(n, m + 1):
-        for i in range(len(string) - j + 1):
-            yield string[i:i + j]
-
-
-def words(string):
-    """
-    Generates all words in the given string.
-
-    Parameters
-    ----------
-    string: str
-        The string from which the words will be extracted.
-
-    Yields
-    ------
-    str
-        A word extracted from the string.
-
-    """
-    for m in re.finditer(r'[A-Za-z0-9]+', string):
-        yield m.group(0)
-
-
-def jaccard_similarity(set0, set1, weights):
-    """
-    Calculates the Jaccard similarity between two sets of terms.
-
-    Parameters
-    ----------
-    set0: set
-        The first set of terms.
-    set1: set
-        The second set of terms.
-    weights: dict
-        A dictionary mapping terms to weights.
-
-    Returns
-    -------
-    float
-        The Jaccard similarity between set0 and set1.
-
-    """
-    intersection = set0 & set1
-
-    if not intersection:
-        return 0
-
-    union = set0 | set1
-    denominator = sum(weights[t] for t in union)
-
-    if not denominator:
-        return 0
-
-    numerator = sum(weights[t] for t in intersection)
-    return numerator / denominator
-
-
-def cosine_similarity(set0, set1, weights):
-    """
-    Calculates the cosine similarity between two sets of terms.
-
-    Parameters
-    ----------
-    set0: set
-        The first set of terms.
-    set1: set
-        The second set of terms.
-    weights: dict
-        A dictionary mapping terms to weights.
-
-    Returns
-    -------
-    float
-        The cosine similarity between set0 and set1.
-
-    """
-    intersection = set0 & set1
-
-    if not intersection:
-        return 0
-
-    length0 = np.sqrt(sum(weights[t]**2 for t in set0))
-    length1 = np.sqrt(sum(weights[t]**2 for t in set1))
-
-    denominator = length0 * length1
-    if not denominator:
-        return 0
-
-    numerator = sum(weights[t]**2 for t in intersection)
-
-    return numerator / denominator
+from .utils import ngrams, nmgrams, words, jaccard_similarity, cosine_similarity
 
 
 class TokenSimilarity():
-    class TokenSimilarity:
-        """
-        Class that implements a token-based similarity model.
-        """
+    """
+    Class that implements a token-based similarity model.
+    """
 
     def __init__(
             self,
